Move messages up to the root level
